# src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: BBYO Connect API
  version: 1.0.0
  summary: Secure closed social network backend for BBYO teens and chapter accounts
  description: |
    Production-oriented backend API for BBYO Connect.

    Security and youth safety features:
    - JWT access tokens + rotating refresh tokens
    - Role-based access control
    - Audit logging
    - Reporting and moderation workflows
    - Language filtering hooks

    Realtime channel (Socket.IO):
    - message.created
    - message.edited
    - message.deleted
    - conversation.updated
    - notification.created
servers:
  - url: http://localhost:4000
    description: Local development
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    StoryIdParam:
      name: storyId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          description: Optional error details
    ApiEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          description: Response payload
        meta:
          description: Pagination or extra metadata
        error:
          oneOf:
            - type: "null"
            - $ref: '#/components/schemas/ApiError'
...
    Locale:
      type: string
      enum: [en, es, he, fr]
...
paths:
...
  /stories:
    get:
      tags: [Stories]
      summary: List active stories
      responses:
        '200':
          description: Stories
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
    post:
      tags: [Stories]
      summary: Create story
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mediaUrl]
              properties:
                mediaUrl:
                  type: string
                  format: uri
                expiresInHours:
                  type: integer
      responses:
        '201':
          description: Story created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
  /stories/{storyId}:
    delete:
      tags: [Stories]
      summary: Delete story
      parameters:
        - $ref: '#/components/parameters/StoryIdParam'
      responses:
        '200':
          description: Story deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
...
  /resources:
    get:
      tags: [Resources]
      summary: Leadership/safety resources
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: language
          in: query
          schema: { $ref: '#/components/schemas/Locale' }
        - name: visibility
          in: query
          schema:
            type: string
            enum: [public, chapter, region, global, private]
      responses:
        '200':
          description: Resource page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
    post:
      tags: [Resources]
      summary: Create resource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, type, language, url, visibility]
              properties:
                title: { type: string }
                type:
                  type: string
                  enum: [article, video, podcast, toolkit, pdf, link]
                language: { $ref: '#/components/schemas/Locale' }
                url: { type: string, format: uri }
                visibility:
                  type: string
                  enum: [public, chapter, region, global, private]
      responses:
        '201':
          description: Resource created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
...